+++ python/GafferUSDTest/USDLayerWriterTest.py
@@ -42,6 +42,7 @@ import pxr.Usd
 import IECore
 import IECoreScene
 
+import Gaffer
 import GafferUSD
 import GafferScene
 import GafferSceneTest
@@ -230,6 +231,20 @@ class USDLayerWriterTest( GafferSceneTest.SceneTestCase ) :
 		reader = GafferScene.SceneReader()
 		reader["fileName"].setValue( compositionFileName )
 		self.assertScenesEqual( reader["out"], spherePrimitive["out"], checks = self.allSceneChecks - { "sets" } )
+
+	def testNoContextLeaks( self ) :
+
+		sphere = GafferScene.Sphere()
+
+		layerWriter = GafferUSD.USDLayerWriter()
+		layerWriter["base"].setInput( sphere["out"] )
+		layerWriter["layer"].setInput( sphere["out"] )
+		layerWriter["fileName"].setValue( self.temporaryDirectory() / "layer.usda" )
+
+		with Gaffer.ContextMonitor( sphere ) as monitor :
+			layerWriter["task"].execute()
+
+		self.assertNotIn( "usdLayerWriter:fileName", monitor.combinedStatistics().variableNames() )
 
 if __name__ == "__main__":
 	unittest.main()